requirements.txt
telethon
pandas
python-dotenv
